Defer sending external memory observations
The previous external memory trace implementation blocked on a fence
after a queue submit, and then immediately sent the contents of the
staging buffer. This is potentially inneficient (waiting for the entire
queue submit to complete); moreover, this approach will deadlock if the
queue submit includes a wait on a timeline semaphore that depends on a
later host-side vkSignalSemaphore call.

This change switches to a deferred strategy--a queue submit containing
an external memory barrier also submits commands to copy the contents
into a staging buffer, and then to set an event; this staging buffer
event is then added to a pending list. At later times (currently, the
end of each vkQueueSubmit call), the events in the pending list are
checked, and, if set, the corresponding staging buffer content is sent.
This means the staging buffer content might be attached to a later
command than the vkQueueSubmit that submitted the corresponding external
memory observation.

This raises the possibility of remaining un-sent external memory
observations at the end of a capture. This requires blocking until the
events in the pending list are set, and then sending the contents. This
is done after each vkQueuePresent call, but should ideally only be done
at the end of the trace. This version ignores timeline semaphores, and
will deadlock if one of these events is blocked on a timeline semaphore
when vkQueuePresent is called.